Use PrincipalWithTypeKeyFormatter in ep role list
This is the principal formatter previously declared for use within
`globus search` commands. It has been expanded in surface area in
order to support the behaviors and differences with
`globus endpoint role list`.
